requirements.txt
torchmetrics~=0.11.4
click~=8.1.3
pyyaml~=6.0
torchvision~=0.14.0
tqdm~=4.65.0
numpy~=1.23.4
albumentations~=1.3.0
pillow~=9.4.0
scikit-learn~=1.2.1
setuptools~=65.6.3
torch~=1.13.1
tensorboardX~=2.6
pylatex~=1.4.1
matplotlib~=3.7.2